Rename package repository -> repositories. (#5)
